Servers now have ports
Added a Server class, which has an InetAddress and a port number. The ClientClass now uses a HashSet<Server> instead of an InetAddress array to store the servers. This allows for different ports to be used for different servers.

Also added the add/remove server functions.

Deleted the ClientTests class, will replace it with something better soon