Separated request sending and generally cleaned up read/write
readMessage and writeMessage functions cleaned up a lot, and their shared request sending code was moved to a separate function.

The readMessage and writeMessage functions now read like psuedocode. It's a bit spaghetti-like, but the shared functions make modification easier, and it's easier to read from a high level perspective